refactor: change structure based on used classes
