+++ src/main/java/webserver/RequestHandler.java
@@ -9,6 +9,8 @@ import java.net.URISyntaxException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.request.Request;
+import webserver.response.Response;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
