+++ src/main/java/webserver/request/QueryParams.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.util.Collections;
 import java.util.Map;
