+++ src/main/java/webserver/request/body/QueryBody.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request.body;
 
 import utils.IOUtils;
 import utils.RequestUtils;
