+++ src/main/java/webserver/request/body/RequestBody.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request.body;
 
 public interface RequestBody {
     String get(String key);
