+++ src/main/java/webserver/response/ContentType.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import java.util.Arrays;
