+++ src/main/java/webserver/response/Response.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.response;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -8,6 +8,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import utils.FileIoUtils;
+import webserver.RequestHandler;
+import webserver.request.Request;
 
 public class Response {
     public static final String EMPTY = "";
