+++ src/test/java/webserver/ControllerTest.java
@@ -7,6 +7,8 @@ import java.io.IOException;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import webserver.request.Request;
+import webserver.response.Response;
 
 class ControllerTest {
