+++ src/test/java/webserver/RequestTest.java
@@ -4,9 +4,12 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.util.Map;
 
 import org.junit.jupiter.api.Test;
+import webserver.request.Method;
+import webserver.request.QueryParams;
+import webserver.request.Request;
+import webserver.request.body.RequestBody;
 
 public class RequestTest {
